+++ crates/cloud-wallet-openid4vc/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod errors;
-pub mod issuance;
+pub mod oid4vci;
+pub mod utils;
