+++ crates/cloud-wallet-openid4vc/src/oid4vci/authz_details.rs
@@ -11,7 +11,7 @@
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-use crate::issuance::claim_path_pointer::ClaimPathPointer;
+use crate::oid4vci::claim_path_pointer::ClaimPathPointer;
 
 /// The type of authorization detail.
 ///
