+++ crates/cloud-wallet-openid4vc/src/oid4vci/authz_server_metadata.rs
