+++ crates/cloud-wallet-openid4vc/src/oid4vci/claim_path_pointer.rs
