+++ crates/cloud-wallet-openid4vc/src/oid4vci/credential_offer.rs
