+++ crates/cloud-wallet-openid4vc/src/oid4vci/css_color.rs
