+++ crates/cloud-wallet-openid4vc/src/oid4vci/error.rs
