+++ crates/cloud-wallet-openid4vc/src/oid4vci/mod.rs
@@ -18,4 +18,3 @@ pub mod notification;
 pub mod query_params;
 pub mod token_request;
 pub mod token_response;
-pub mod utils;
