+++ crates/cloud-wallet-openid4vc/src/utils/mod.rs
