+++ tests/consent.rs
@@ -2,15 +2,15 @@ use cloud_identity_wallet::{
     domain::models::issuance::FlowType,
     session::{IssuanceSession, IssuanceState, MemorySession, SessionStore},
 };
-use cloud_wallet_openid4vc::issuance::client::ResolvedOfferContext;
+use cloud_wallet_openid4vc::oid4vci::client::ResolvedOfferContext;
 use serde_json::json;
 use std::time::Duration;
 
 fn create_test_session(flow: FlowType) -> IssuanceSession {
-    use cloud_wallet_openid4vc::issuance::authz_server_metadata::AuthorizationServerMetadata;
-    use cloud_wallet_openid4vc::issuance::client::IssuanceFlow;
-    use cloud_wallet_openid4vc::issuance::credential_offer::CredentialOffer;
-    use cloud_wallet_openid4vc::issuance::issuer_metadata::CredentialIssuerMetadata;
+    use cloud_wallet_openid4vc::oid4vci::authz_server_metadata::AuthorizationServerMetadata;
+    use cloud_wallet_openid4vc::oid4vci::client::IssuanceFlow;
+    use cloud_wallet_openid4vc::oid4vci::credential_offer::CredentialOffer;
+    use cloud_wallet_openid4vc::oid4vci::issuer_metadata::CredentialIssuerMetadata;
 
     let offer: CredentialOffer = serde_json::from_value(json!({
         "credential_issuer": "https://issuer.example.com",
