+++ tests/issuance_tx_code.rs
@@ -19,7 +19,7 @@ use cloud_identity_wallet::{
     server::Server,
     session::{IssuanceSession, IssuanceState, MemorySession, SessionStore},
 };
-use cloud_wallet_openid4vc::issuance::client::{Config as Oid4vciClientConfig, Oid4vciClient};
+use cloud_wallet_openid4vc::oid4vci::client::{Config as Oid4vciClientConfig, Oid4vciClient};
 use reqwest::Client;
 use url::Url;
 use uuid::Uuid;
